+++ app/javascript/components/Home.jsx
@@ -30,58 +30,9 @@ export default(props) => {
 		}
 	];
 
-	const MOCK_ORDERS = [
-		{
-			id: 1,
-			user: {
-				username: "Iris"
-			},
-			items: [
-				{
-					id: 2,
-					name: "Signature Ddough",
-					quantity: 3,
-					price: 1.2
-				},
-				{
-					id: 3,
-					name: "Bbough",
-					quantity: 3,
-					price: 1.5
-				}
-			]
-		},
-		{
-			id: 2,
-			user: {
-				username: "Alan"
-			},
-			items: [
-				{
-					id: 3,
-					name: "Bbough",
-					quantity: 5,
-					price: 1.5
-				}
-			],
-		},
-		{
-			id: 3,
-			username: "Andrew",
-			items: [
-				{
-					id: 1,
-					name: "Classic Donut",
-					quantity: 1,
-					price: 1.0
-				}
-			],
-		}
-	];
-
 	return (
 		<div className="store-container">
-			<OrderList orders={MOCK_ORDERS} />
+			{props.role == "seller" && (<OrderList />)}
 			<StoreDisplay itemList={MOCK_ITEMS} role={props.role} />
 		</div>
 	);
